requirements.txt
flask
mysql-connector-python
pandas
python-dotenv
bcrypt
